# Optimize InefficientlyConvertToJS