--- a/go.mod
+++ b/go.mod
@@ -61,7 +61,6 @@ require (
 	github.com/mattn/go-sqlite3 v2.0.3+incompatible
 	github.com/multiformats/go-multiaddr v0.2.1
 	github.com/multiformats/go-multiaddr-dns v0.2.0
-	github.com/ocdogan/rbt v0.0.0-20160425054511-de6e2b48be33
 	github.com/pborman/uuid v0.0.0-20180906182336-adf5a7427709
 	github.com/plaid/go-envvar v1.1.0
 	github.com/prometheus/tsdb v0.10.0 // indirect
@@ -70,7 +69,6 @@ require (
 	github.com/sirupsen/logrus v1.4.2
 	github.com/status-im/keycard-go v0.0.0-20190424133014-d95853db0f48 // indirect
 	github.com/stretchr/testify v1.4.0
-	github.com/syndtr/goleveldb v1.0.1-0.20190923125748-758128399b1d
 	github.com/tyler-smith/go-bip39 v1.0.2 // indirect
 	github.com/wsxiaoys/terminal v0.0.0-20160513160801-0940f3fc43a0 // indirect
 	github.com/xeipuuv/gojsonpointer v0.0.0-20190809123943-df4f5c81cb3b // indirect