--- a/packages/browser/go/jsutil/jsutil_bench_test.go
+++ b/packages/browser/go/jsutil/jsutil_bench_test.go
@@ -1,3 +1,5 @@
+// +build js,wasm
+
 package jsutil
 
 import (