This PR optimizes `jsutil.InefficientlyConvertToJS` by using `JSON.parse` in JavaScript land.

I also added some benchmarks, which converts slices of `testValue` with lengths between 1 and 1,000. The benchmarks show up to a 10x performance increase for large values.

For future optimizations, we should look at manually writing [`JSValue`](https://golang.org/pkg/syscall/js/#Wrapper) or [`MarshalJSON`](https://golang.org/pkg/encoding/json/#Marshaler) methods for types that we need to frequently convert in hot code paths.

Here are the results with the old version:

```
> WASM_INIT_FILE="$(pwd)/packages/test-wasm/dist/browser_shim.js" GOOS=js GOARCH=wasm go test -exec="wasmbrowsertest" ./packages/browser/go/jsutil -bench=. -run=NONE
goos: js
goarch: wasm
pkg: github.com/0xProject/0x-mesh/packages/browser/go/jsutil
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_1         	   10683	    100016 ns/op
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_10        	    1705	    737314 ns/op
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_100       	     172	   7050261 ns/op
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_1000      	      15	  89819682 ns/op
```

And here are the results with the new version:

```
> WASM_INIT_FILE="$(pwd)/packages/test-wasm/dist/browser_shim.js" GOOS=js GOARCH=wasm go test -exec="wasmbrowsertest" ./packages/browser/go/jsutil -bench=. -run=NONE
goos: js
goarch: wasm
pkg: github.com/0xProject/0x-mesh/packages/browser/go/jsutil
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_1         	   36669	     31170 ns/op
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_10        	    9052	    118822 ns/op
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_100       	    1170	    933889 ns/op
BenchmarkInefficientlyConvertToJS/BenchmarkInefficientlyConvertToJS_1000      	     130	   9021154 ns/op
```

|  Size  | Old         | New        | Change  |
|-------|-------------|------------|---------|
| 1     | 100,016 ns   | 31,170 ns   | -68.8% |
| 10    | 737,314 ns   | 118,822 ns  | -83.9% |
| 100   | 7,050,261 ns  | 933,889 ns  | -86.7% |
| 1,000 | 89,819,682 ns | 9,021,154 ns | -89.9% |