@pgebheim 2020-06-10
Lovely example of micro optimizations when they're needed.

Thanks guys